Created addUser function for signup. Correctly adds information to sql database.
